Include compatible runtimes in ExpressionData
