+++ src/Language/Mimsa/Backend/Types.hs
@@ -13,6 +13,7 @@ import Language.Mimsa.Types.Store
 type BackendM ann = Either (BackendError ann)
 
 data Backend = CommonJS
+  deriving (Eq, Ord, Show)
 
 data Renderer ann a = Renderer
   { renderFunc :: Name -> Expr Name ann -> BackendM ann a,
