+++ src/Language/Mimsa/Server/Handlers.hs
@@ -6,11 +6,8 @@
 
 module Language.Mimsa.Server.Handlers
   ( ProjectData (..),
-    UnitTestData (..),
-    ExpressionData (..),
     fromActionM,
     projectDataHandler,
-    expressionDataHandler,
     loadProjectHandler,
     evaluateTextHandler,
     createNewUnitTestsHandler,
@@ -24,7 +21,6 @@ module Language.Mimsa.Server.Handlers
     readStoreHandler,
     writeStoreHandler,
     createUnitTestHandler,
-    mkUnitTestData,
   )
 where
 
@@ -32,11 +28,8 @@ import qualified Control.Concurrent.STM as STM
 import Control.Monad.Except
 import qualified Data.Aeson as JSON
 import Data.Bifunctor (first)
-import Data.Coerce
 import Data.Foldable (traverse_)
 import Data.Map (Map)
-import qualified Data.Map as M
-import qualified Data.Set as S
 import Data.Swagger
 import Data.Text (Text)
 import GHC.Generics
@@ -62,7 +55,6 @@ import Language.Mimsa.Types.ResolvedExpression
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
 import Language.Mimsa.Types.Swaps
-import Language.Mimsa.Types.Typechecker
 import Servant
 
 -----
@@ -108,22 +100,6 @@ data ProjectData = ProjectData
   }
   deriving (Eq, Ord, Show, Generic, JSON.ToJSON, ToSchema)
 
-data UnitTestData = UnitTestData
-  { utdTestName :: Text,
-    utdTestSuccess :: Bool,
-    utdBindings :: Map Name Text
-  }
-  deriving (Eq, Ord, Show, Generic, JSON.ToJSON, ToSchema)
-
-mkUnitTestData :: Project ann -> UnitTest -> UnitTestData
-mkUnitTestData project unitTest = do
-  let getDep = (`findBindingNameForExprHash` project)
-  let depMap = mconcat (getDep <$> S.toList (utDeps unitTest))
-  UnitTestData
-    (coerce $ utName unitTest)
-    (coerce $ utSuccess unitTest)
-    (coerce <$> depMap)
-
 -- read the store from mutable var to stop repeated loading of exprs
 readStoreHandler :: MimsaEnvironment -> Handler (Store Annotation)
 readStoreHandler mimsaEnv = do
@@ -146,36 +122,6 @@ projectDataHandler mimsaEnv env = do
       projHash
       (outputBindings env)
       (outputTypeBindings env)
-
-data ExpressionData = ExpressionData
-  { edHash :: Text,
-    edPretty :: Text,
-    edType :: Text,
-    edBindings :: Map Name Text,
-    edTypeBindings :: Map TyCon Text,
-    edUnitTests :: [UnitTestData]
-  }
-  deriving (Eq, Ord, Show, Generic, JSON.ToJSON, ToSchema)
-
-expressionDataHandler ::
-  Project Annotation ->
-  StoreExpression Annotation ->
-  MonoType ->
-  Handler ExpressionData
-expressionDataHandler project se mt = do
-  let exprHash = getStoreExpressionHash se
-      tests =
-        mkUnitTestData project
-          <$> M.elems
-            (getTestsForExprHash project exprHash)
-  pure $
-    ExpressionData
-      (prettyPrint exprHash)
-      (prettyPrint (storeExpression se))
-      (prettyPrint mt)
-      (prettyPrint <$> getBindings (storeBindings se))
-      (prettyPrint <$> getTypeBindings (storeTypeBindings se))
-      tests
 
 -- given a project hash, find the project
 loadProjectHandler ::
