+++ src/Language/Mimsa/Server/Project/AddUnitTest.hs
@@ -15,6 +15,7 @@ import Data.Swagger
 import Data.Text (Text)
 import GHC.Generics
 import qualified Language.Mimsa.Actions.AddUnitTest as Actions
+import Language.Mimsa.Server.ExpressionData
 import Language.Mimsa.Server.Handlers
 import Language.Mimsa.Server.Types
 import Language.Mimsa.Types.Project
